Ensure stored procedure names are not controllable
